Made sprint bar fade in and out on use.
The sprint bar now fades in for 1 second when sprinting.  It then waits until 3 seconds after the player stops sprinting AND stamina has reached it's max again to fade back out.  If the player sprints again as it is fading out, it will fade back in from that exact opacity level.

Everything should be working smoothly here, but if anything comes up, let me know.  All of these numbers can be changed if we decide they should be something else.  If we decide to change the amount of stamina the player has, I will need to make adjustments to this system.  We may want to change the numbers in the original stamina system to make the bar move smoother when sprinting and resting.

Thank you!