Added Simple Crouch
Smoothly lowers and raises camera with left control.  It also smoothly changes the half height of the capsule component.

We will probably have to make some changes to the numbers and the way this works to make it look better and to make it useful.  We will also have to make it work with animations which I am not sure how we will do yet.